+++ ledger-tool/src/args.rs
@@ -44,7 +44,6 @@ pub fn accounts_db_args<'a, 'b>() -> Box<[Arg<'a, 'b>]> {
             .value_name("PATH")
             .takes_value(true)
             .multiple(true)
-            .requires("enable_accounts_disk_index")
             .help(
                 "Persistent accounts-index location. May be specified multiple times. [default: \
                  <LEDGER>/accounts_index]",
@@ -62,12 +61,29 @@ pub fn accounts_db_args<'a, 'b>() -> Box<[Arg<'a, 'b>]> {
             .takes_value(true)
             .help("Pre-allocate the accounts index, assuming this many accounts")
             .hidden(hidden_unless_forced()),
-        Arg::with_name("enable_accounts_disk_index")
-            .long("enable-accounts-disk-index")
-            .help("Enables the disk-based accounts index")
+        Arg::with_name("accounts_index_limit")
+            .long("accounts-index-limit")
+            .value_name("VALUE")
+            .takes_value(true)
+            .possible_values(&[
+                "minimal",
+                "25GB",
+                "50GB",
+                "100GB",
+                "200GB",
+                "400GB",
+                "800GB",
+                "unlimited",
+            ])
+            .default_value("unlimited")
+            .help("Sets the memory limit for the accounts index")
             .long_help(
-                "Enables the disk-based accounts index. Reduce the memory footprint of the \
-                 accounts index at the cost of index performance.",
+                "Sets the memory limit for the accounts index. The size options will limit the \
+                 accounts index memory to the specified value. E.g. \"50GB\" means the accounts \
+                 index may use up to 50 GB of memory. The \"unlimited\" option keeps the entire \
+                 accounts index in memory. The \"minimal\" option reduces memory usage as much as \
+                 possible. All index entries that are not in memory are kept in the disk-backed \
+                 index. The disk-backed index has lower performance; prefer higher limits here.",
             ),
         Arg::with_name("accounts_db_skip_shrink")
             .long("accounts-db-skip-shrink")
@@ -253,10 +269,21 @@ pub fn get_accounts_db_config(
     let accounts_index_bins = value_t!(arg_matches, "accounts_index_bins", usize).ok();
     let num_initial_accounts =
         value_t!(arg_matches, "accounts_index_initial_accounts_count", usize).ok();
-    let accounts_index_index_limit = if !arg_matches.is_present("enable_accounts_disk_index") {
-        IndexLimit::InMemOnly
-    } else {
-        IndexLimit::Minimal
+    let accounts_index_limit =
+        value_t!(arg_matches, "accounts_index_limit", String).unwrap_or_else(|err| err.exit());
+    let index_limit = match accounts_index_limit.as_str() {
+        "minimal" => IndexLimit::Minimal,
+        "25GB" => IndexLimit::Threshold(25_000_000_000),
+        "50GB" => IndexLimit::Threshold(50_000_000_000),
+        "100GB" => IndexLimit::Threshold(100_000_000_000),
+        "200GB" => IndexLimit::Threshold(200_000_000_000),
+        "400GB" => IndexLimit::Threshold(400_000_000_000),
+        "800GB" => IndexLimit::Threshold(800_000_000_000),
+        "unlimited" => IndexLimit::InMemOnly,
+        x => {
+            // clap will enforce only the above values are possible
+            unreachable!("invalid value given to `--accounts-index-limit`: '{x}'")
+        }
     };
     let accounts_index_drives = values_t!(arg_matches, "accounts_index_path", String)
         .ok()
@@ -265,7 +292,7 @@ pub fn get_accounts_db_config(
     let accounts_index_config = AccountsIndexConfig {
         bins: accounts_index_bins,
         num_initial_accounts,
-        index_limit: accounts_index_index_limit,
+        index_limit,
         drives: Some(accounts_index_drives),
         ..AccountsIndexConfig::default()
     };